requirements.txt
decorator==4.4.2
easydict==1.7
fvcore==0.1.3.post20210311
imageio==2.9.0
iopath==0.1.4
munkres==1.1.4
networkx==2.5
numpy==1.17.3
opencv-python==4.4.0.44
portalocker==2.2.1
progress==1.5
pthflops==0.4.0
pywavelets==1.1.1
scikit-image==0.17.2
tabulate==0.8.9
thop
tifffile==2020.10.1
timm==0.3.4
torchscan==0.1.1
torchstat==0.0.7
torchvision==0.8.2
einops==0.3.0
scikit-learn
